Move VariableIterator to builtins
